Move ast out of directory
